+++ test/fixtures/config/webpack.config.js
@@ -6,9 +6,9 @@
  * Example usage (make sure to be in project root):
  *
  * ```sh
- * $ export WEBPACK_VERSION=4; \
+ * $ export WEBPACK_VERSION=5; \
  *   WEBPACK_MODE=development \
- *   WEBPACK_CWD=../../test/fixtures/duplicates-esm \
+ *   WEBPACK_CWD=../../test/fixtures/hidden-app-roots \
  *   NODE_PATH="${PWD}/node_modules/webpack${WEBPACK_VERSION}/node_modules:${PWD}/node_modules" \
  *   node test/fixtures/packages/webpack.js \
  *     --config ../../test/fixtures/config/webpack.config.js
@@ -39,6 +39,7 @@ const vers = process.env.WEBPACK_VERSION;
 if (!vers) {
   throw new Error("WEBPACK_VERSION is required");
 }
+const versNum = parseInt(vers, 10);
 
 const cwd = process.env.WEBPACK_CWD;
 if (!cwd) {
@@ -46,7 +47,9 @@ if (!cwd) {
 }
 
 const outputPath = resolve(cwd, `dist-${mode}-${vers}`);
-const webpack4 = {
+
+// Webpack 4+
+let configModern = {
   mode,
   devtool: false,
   context: resolve(cwd),
@@ -72,7 +75,10 @@ const webpack4 = {
   },
   plugins: [
     new StatsWriterPlugin({
-      fields: ["assets", "modules"]
+      fields: ["assets", "modules"],
+      stats: {
+        source: true // Needed for webpack5+
+      }
     }),
     DuplicatesPlugin ? new DuplicatesPlugin({
       verbose: true,
@@ -81,36 +87,39 @@ const webpack4 = {
   ].filter(Boolean)
 };
 
-const webpack1Module = {
-  loaders: webpack4.module.rules.map((rule) => ({
-    test: rule.test,
-    loader: rule.use
-  }))
-};
-
-const webpackOld = {
-  devtool: webpack4.devtool,
-  context: webpack4.context,
-  entry: webpack4.entry,
-  output: webpack4.output,
-  // TODO(66): Add minify thing here -- mode === "production",
-  // https://github.com/FormidableLabs/inspectpack/issues/66
-  module: vers === "1" ? webpack1Module : webpack4.module,
-  plugins: webpack4.plugins
-};
-
-// Choose appropriate version.
-let config = vers === "4" ? webpack4 : webpackOld;
-
 // Dynamically try to import a custom override from `CWD/webpack.config.js`
 try {
   const webpack = require(`webpack${vers}/lib`); // eslint-disable-line global-require
   const override = require(resolve(cwd, "webpack.config.js")); // eslint-disable-line global-require
-  config = override(webpack, config);
+  configModern = override(webpack, configModern, versNum);
 } catch (err) {
   if (err.code !== "MODULE_NOT_FOUND") {
     throw err;
   }
 }
+
+const webpack1Module = {
+  loaders: configModern.module.rules.map((rule) => ({
+    test: rule.test,
+    loader: rule.use
+  }))
+};
+
+// Webpack 2-3
+const configLegacy = {
+  devtool: configModern.devtool,
+  context: configModern.context,
+  entry: configModern.entry,
+  output: configModern.output,
+  // TODO(66): Add minify thing here -- mode === "production",
+  // https://github.com/FormidableLabs/inspectpack/issues/66
+  module: versNum === 1 ? webpack1Module : configModern.module,
+  plugins: configModern.plugins,
+  resolve: configModern.resolve
+};
+
+// Choose appropriate version.
+// eslint-disable-next-line no-magic-numbers
+const config = versNum >= 4 ? configModern : configLegacy;
 
 module.exports = config;