+++ test/fixtures/loaders/src/index.js
@@ -1,22 +1,15 @@
 
 /* eslint-disable no-console */
-/* globals global */
 
 import text from "./hello.txt";
 import style from "./style.css";
 
-// Use expose loader to make global
-require("expose-loader?BunBun!./bunny"); // eslint-disable-line import/no-unresolved
+// Legacy: just require a file (gave up on `expose-loader` global in webpack5
+// upgrade).
+require("./bunny");
 
 const hello = () => "hello world";
 
 console.log("hello", hello());
 console.log("text", text);
 console.log("style", style.toString());
-
-let root = typeof window !== "undefined" && window;
-if (!root && typeof global !== "undefined") {
-  root = global;
-}
-
-console.log("global", root.BunBun);